fix : updatePlat(int $id)  l'upload d'une image ne vérifiait pas si le fichier existe , erreur à l'upload . Css vue menu détaillé en cours , import imgaes diverses pour alimenter les plats et menus

+++ app/controllers/PlatController.php
@@ -219,7 +219,7 @@ class PlatController{
                 header('Location: /plats/edit/' . $id .'?error=' . urlencode($error));
                 exit();
             }
-            if($plat['chemin_photo']){
+            if($plat['chemin_photo'] && file_exists(__DIR__ . "/../../public/" . $plat['chemin_photo'])){
                 unlink(__DIR__ . "/../../public/" . $plat['chemin_photo']);
             }
